Fixed issue with cached host viewport state.
